Refactor jwt_knox so that it forms a ViewSet
Now the ViewSet needs a different authentication_classes depending on
whether we are using get_token or the rest of the ViewSet functions.
Because of that, we need to implement get_authenticators.

Instead of implementing get_authenticators manually, as it will be
called by super().initialize_request(request, *args, **kwargs), we
implement a PerViewAuthentication mixin instead, that, hopefully, we
will be able to reuse elsewhere.

This way, we re-call initialize_request in initial(), after the
ViewSet's request has already been initialized and thus we are able to
use self.action in the subclassed per-view authenticator logic (and
users of the mixin needn't be aware on the ViewSet initialization
lifecycle).