JWT-Knox tests: Use .force_authenticate; no .login
Client.login() is a Django Client function which leverages
SessionAuthentication, and its lifecycle seems not to be compatible with
our custom mixin's patching. So, instead of using SessionAuthentication
we can either use BasicAuthentication (which seems not to suffer from
this issue) or using rest_framework's APIClient.force_authenticate,
which understands rest_framework's Request object and forces
authentication on it, instead of relying on a Django Session.

This change forces us to cache self.user (the created user model) to
provide it to the request.